Bring back whitespace tolerance in selective `phpcs:ignore` comments
To selectively ignore /disable/enable standards, sniff categories, sniffs or error codes, the `phpcs:ignore`, `phpcs:disable` and `phpcs:enable` annotations take a comma-separated list of rule references and allow for an optional `-- comment` at the end of the annotation.

While not strictly speaking "officially supported", the comma-separated list was whitespace tolerant in PHPCS 3.x.

This was unintentionally broken in PR 123, but only for the `phpcs:ignore` syntax. The `phpcs:disable`/`phpcs:enable` annotations were not affected.

This commit fixes this bug and brings back whitespace tolerance in the comma-separated list of rule references for `phpcs:ignore`.

It also adds a set of tests for all three types of annotations to safeguard this.
The addition of these tests effectively means that this whitespace tolerance is now officially supported.

Fixes 1277