Define polyfilled constants from Tokenizer as int
The value of these constants are not stable, and therefore already cannot be
relied upon. This is because the specific values that PHP assigns can change
with different versions of PHP. PHPCS does not use the values of these
constants (other than to look up their name using the Tokens::tokenName()
method).

There are other tools which also polyfill these constants. Some of those tools
also perform validation on the value for these constants. In order to play
nicely with the arbitrary validation that other tools perform on these
constants, we are switching from string values to integer values.

All PHPCS 'native' tokens currently have reliable values. In line with PHP T_*
constants, the values of these tokens should never be relied upon. In a future
version of PHPCS, the values for these tokens will switch from strings to
integers.

Existing tests already cover the use of these constants and do not require
adjustment for the code being changed here.