E2E tests: use new bashunit assertion
PR 1296 added a new test which _should_ also verify that the command had a non-zero exit code, but couldn't do so at the time as this wasn't (yet) supported by Bashunit.

As of Bashunit 0.26.0, a new `assert_unsuccessful_code` is available which fixes that.

This commit starts using that assertion.

Refs:
* https://github.com/TypedDevs/bashunit/releases/tag/0.26.0
* TypedDevs/bashunit 505
* TypedDevs/bashunit 506