Tokenizer/PHP: PHP close tag should be regarded as scope opener for switch case/default statements
In PHP, if the statement before a PHP close tag is "unclosed", PHP will automagically insert a semi-colon before the PHP close tag when compiling the file.

As `case` and `default` condition statements in a `switch` structure can be "closed" with a semi-colon, this will not result in a parse error, but will be perfectly valid code.

However, the practice of "closing" `case`/`default` statements with a semi-colon is deprecated as of PHP 8.5 and while looking into that deprecation, I realized that the deprecation also applies to the "implied"/automagically inserted semi-colon when a `case`/`default` statement ends on a PHP close tag.

Digging deeper, it turned out that a PHP close tag was not supported in PHPCS as a scope opener for `case`/`default` statements, which in turn meant that if that type of code was seen by sniffs, all sorts of false positives/negatives and potential `Undefined array key "scope_opener/closer"` notices would start to get thrown.

This commit fixes the tokenizer part of the issue and adds tests with a variety of scope openers for `case`/`default` statements to a number of Tokenizer related test files.