PSR2/SwitchDeclaration: make implied semicolon via close tag auto-fixable
The `PSR2.ControlStructures.SwitchDeclaration` sniff includes a rule/error flagging when the scope opener is not a `:` colon.
This error was previously already made auto-fixable when the `scope_opener` is a semi-colon (via PR 1161).

This commit now also makes that error auto-fixable if the `scope_opener` is a PHP close tag.

Includes tests.
These tests also document how the sniff otherwise behaves when the code within a `switch` moves in and out of PHP a lot.