Squiz/SwitchDeclaration: make implied semicolon via close tag auto-fixable
The `Squiz.ControlStructures.SwitchDeclaration` sniff does _not_ include a rule/error for the scope opener not being a `:` colon, though all the error messages have the presumption of the scope opener being a colon in the textual content.

The sniff does have tests with a semicolon scope opener and as there is no error related to the semicolon being a scope opener, I can only presume that it was previously decided to allow that in the Squiz standard.
Considering that this is a code style sniff, it is not for this sniff to enforce PHP cross-version compatibility of code, so the PHP 8.5 deprecation is irrelevant and there is no reason to change the behaviour of the sniff for semicolon scope openers.

However, the sniff definitely did not function correctly for code moving in and out of PHP within a `switch` statement.
PR 1315 was a first step towards supporting this.

This commit now adds a new error to disallow a PHP close tag as the "scope_opener" for `case`/`default` statements.
The error is auto-fixable and will insert a colon before the PHP close tag as the fix.
With this additional error in place, the sniff's handling of code within a `switch` moving in and out of PHP has greatly improved.

Includes tests.