GH Actions: split "validate" workflow
GitHub has the annoying habit of disabling workflows with a cron job after two months if the repo doesn't see any activity.

While this won't happen easily for this repo, it may happen in forks and this creates the following problem:
* If the same workflow is used for both the cron job as well as the push/pull_request CI checks...
* ... and a repo (fork) doesn't have any activity in two months time...
* ... the workflow gets disabled...
* ... which then also means that CI checks will no longer be run on newly pushed branches....
* ... which means that contributors don't get feedback ahead of creating a PR, increasing the likelyhood the PR will fail CI.

Looking at this workflow critically, only the markdown checks should be run via a cron job. All other jobs should only run on push/pull.

With that in mind, this commit splits the workflow into two workflows:
* One workflow will be triggered via `cron` and will only run the markdown related checks.
* One workflow will have all the other triggers (`push`/`pull_request`/`workflow_dispatch`) and will run the full set of validation checks.

This way, if the cron job workflow gets disabled, the workflow which is used for the other triggers will continue to function.